BindableBaseUtil不管用，换为BindableBase
